Refactor Dagger element group name logic
Logic to determine related Dagger element group/category names is
currently duplicated in two different places, and breaks the design
principle of having the "concepts" understand everything about a
particular usage. Additionally, splitting the logic like this causes us
to evaluate some of the same things multiple times. Figuring out how two
subcomponents are related, for example, depends on their structure in
the Dagger graph. We know that already when we identify them, but if we
need to figure out those category names separately later, we would need
to parse that graph relationship a second time.

This change modifies the "get related elements" logic to return a group
name along with the related element. This has the side effect of
allowing us to remove the UsageTypeProvider extension completely, since
if we now can return an object implementing `UsageWithType`, the
platform no longer needs to call us back later to figure out what the
usage type is.

Additionally, this change adds a few types of related items to
Components/Subcomponents/Modules that do not come from the index. Now
that the group names can be specified directly, we're able to return
those without having to worry about them being correctly identified
later.

Bug: 265846405
Test: Included
Change-Id: Idd3185f0492ef21082a78c99e08fc15106ce9720