Extract IDE classpath from product-info.json
Unfortunately, the IDE runs with a hardcoded list of platform jars
on the classpath, rather than all jars under the lib directory.

To match this behavior in tests, we must extract the list of classpath
jars from product-info.json (just like gradle-intellij-plugin does).
Otherwise we run the risk of getting NoClassDefFoundError in
production (as in b/265231036).

The most immediate impact of this change is that testFramework.jar
will be excluded from the production classpath (studio-sdk).

Bug: 265335760
Test: existing Bazel tests after running update_sdk.py
Test: full JPS build after running update_sdk.py
Change-Id: I706447510c38238823cf5aceb99378e314e266d3