Extract startTracing logic to util function
The CpuProfilerStage's startCapture method
contains logic to execute a start trace command
and register listeners for the expected response
events. This logic, as done with stopTracing in
CpuProfiler (used in CpuProfilerStage's stopCapture
method), can be extracted to be a utility method.

This will help with code reuse as the Task-Based Framework
will likely require access to this functionality.

Bug: 270211065
Test: existing
Change-Id: Ia8fd87560f8f53eb89902eb4799240cc0a8ba2fb