Initialize Gradle local java.home when the field isn't defined
Given a project or branch that uses the macro #GRADLE_LOCAL_JAVA_HOME
in case user trigger sync and they haven't defined java.home, the new
InitializeGradleLocalJavaHomeListener will initialize this for
their project, in order to improve the experience for those projects
that are using the new macro.

This will apply for each project gradle root and the selection
criteria for the JDK path will follow:
 1). The project JDK defined on .idea/misc.xml if valid
 2). The default JDK configured under project structure if valid
 3). The embedded JDK

Knowing that project JDK on the lastest changes represents the
gradle JDK after a successful sync, this will result in most cases
to doesn't have any change in jdk path used to trigger gradle sync.

Bug: 214428183
Test: GradleConfigManagerTest
Test: SingleGradleRootSyncUseGradleLocalJavaHomeIntegrationTest
Test: MultipleGradleRootSyncUseGradleLocalJavaHomeIntegrationTest
Change-Id: I488e8d33d6b22ec56b2be11b40b802297a709e52