Add support to multiple gradle roots project to existing hyperlinks
Given multiple gradle roots, when display the project gradle
settings we will pre-select the current gradle project for the
user, opening directly the linked configuration.

The AndroidProjectSettingsService.chooseJdkLocation allows to specify
current gradle root path, that will be used to select the project.
In those cases when the provided path is null or not valid the first
project will be selected by default.

Bug: 214428183
Test: SelectJdkFromFileSystemHyperlinkTest
Test: OpenProjectJdkLocationListenerTest
Change-Id: Ia36e94ffebe27d0732cb6ec35824333aafbf0860