Rework variant conflict display
Build variant view will now show variant conflicts as warning instead of
error. We are now going to display the warning on the dependent
module(consumer) rather than dependency. Since the variant conflict
cannot be reliably resolved, we are not going to attempt to resolve it
automatically but instead present the details of conflict in a tooltip
for user to manually take an action if they desire to.

Bug: 201824664
Test: ModuleVariantConflictTest
Change-Id: I0098ea5434285f52511c457da800284572ce7444