adblib: Add "boost" support for JDWP processes
Given the `WAIT` packet behavior (see b/271572555 for
context) and the goal to eventually allow multiple instances
of Android Studio (or IntelliJ + Android plugin) running
concurrently when acquiring JDWP sessions on Android
device, this change implements "boosting" JDWP process
discovery when the Android Studio application is activated,
and "un-boosting" when de-activated.

This ensures the currently active Android Studio instance
will always have priority accessing the JDWP session
of new processes, which is critical for debugging (and
useful for other use cases, such are running an app).

The goal is to make sure we get to a state where having
multiple instances of Studio (or IJ+Android plugin)
running at the same time "just works" (*) from an
end-user perspective.

(*) "just works": most of the time, given that we
can't rely build a fool proof solution given the current
limitations of the Android Debugger and the way JDWP
sessions are contrained.

Bug: 271572555
Test: Included
Change-Id: I899bd8dafd0f5fb6e9c6e883aad7bc2aae5d7564