Update text for gradle task list setting
When SKIP_GRADLE_TASKS_LIST is true we display a message saying that the
Gradle task list is not build, but now it is partially build to contain
test tasks.

This change updates the message to:

"Only test tasks are configured during Gradle Sync. Configuring all tasks
can impact Gradle sync performance on large projects."

and also updates the Gradle setting checkbox text to:

"Configure all Gradle tasks during Gradle Sync (this can make Gradle Sync
slower)"

Note that this change reverses the meaning of having this checkbox
selected. Because of this and in order to keep user settings as they are,
the form that shows this information has to reverse the values read and
written to SKIP_GRADLE_TASKS_LIST.

Fixes: 243423845
Test: ExperimentalSettingsConfigurableTest
Change-Id: Ie7c8bb727a2f499bd9f15cb998d716ab272913fa