Set minimum height on process selection popup
When importing a trace that contains multi-process data,
the user is prompted with a popup dialog to select a
process. When the height of this dialog is too small,
it removes/hides the process selection dropdown. To
prevent this from happening, a minimum height on the
popup dialog has been set.

Fixes: 186579980
Test: manual
Change-Id: I8438fe43e726afe84b0a3ebf77b5bbb996ff6190