Add adt-ui-compose module + sample
A module for Compose UI has been created. It was and is
initially intended for the Profiler tool, but has been
generalized for other Android Dev Tools to contribute or
take from, hence the module name "adt-ui-compose". It
contains the necessary dependencies for creating compose
UI within the src of the module. These composables can
be used by modules that depend on this module.

The "compose-desktop-runtime" module-library was exposed
as an export. This was done on purpose so that modules
that depend on this module are allowed to take advantage
of apis that work in conjunction with Compose UI, such
as "MutableState".

A sample compose app was added to test out the dependencies.
This was a sample mostly copied from the compose-samples module.

Bug: 272822954
Test: none
Change-Id: Ie85e2461773d89fc1cd87c2420a6a5ee776f0579