Keep first capitalization of argument name for SafeArgs
SafeArgs automatically generate classes based on the navigation XML
file. If generating Kotlin classes, it translates argument names defined
in XML into method argument names for the Kotlin methods. The
transformation applied to the argument names is a camel case one, but it
keeps to capitalization of the first letter as is.
This makes sure that the capitalization of the first letter is indeed
preserved.

Fixes: 204522151
Test: test updated
Change-Id: I6b14e8889a78ea66f2e3765d738a9c4750c9dd80