Move MaterialVdIconsProvider to dedicated module
This logic will be reused in the Compose module, so by pulling it out
separately it can be referenced from there.

All of the testResources files are copied directly from the android-npw module. Some do not show up as moves because I had to also leave the original file in android-npw for other tests.

Bug: 272127130
Test: included
Change-Id: I0ffefc8a942d3a05152e9b432eb0a7b2f40e4125