Change the keymap of zoom-to-actual
Before this CL, the shortcut of zoom-to-actual is "CTRL 0" (or "CMD 0"
on OSX).

Replace it with "CTRL ." (or "CMD ." on OSX) because it is duplicated
with Commit tool window.

NUMPAD0 is considered as another key, and Commit tool window doesn't
use it as the shortcut. Thus, the "CTRL NUMPAD0" (or "CMD NUMPAD0" on
OSX) is kept for back compatible when zooming by NUMPAD.

Fixes: 262769449
Test: N/A
Change-Id: I0eab1d2a1e3407a0349cfdc04df194572b0e0ab1