Add a StudioFlag to skip building runtime classpaths
This flag will be used to turn on the experimental feature that tried to
save project import time be not resolving the runtime classpath for
libraries. Instead we will obtain the classpath from the apps runtime
dependency graph.

This flag currently defaults to off which should have no effect on the
behaviour of studio unless turned on.

Bug: 267066187
Test: N/A
Change-Id: If2eb7f06f004be2e09718db75198ee8e1832e810