Add DependencyRef with lookup function to IdeDependenciesCore
This will be used to allow us to have different types of
IdeDepdnenciesCore that can refer to other instances. We add
IdeDependenciesCoreRef although this is currently unsed we include for
context.

We also add a small wrapper class to make it clear what these Int values
refer to and to allow us to absract the details from callers.

Bug: 267066187
Test: existing
Change-Id: Ib648b596d09519670c710e5da821f6f2a325495a