Configure new created projects with Embedded JDK by default
Right now, the configuration of the new projects relay completely on
IdeSdks.sdk() method. This method ends up searching in the jdk.table
returning the first JDK compatible version, making this resolution
inconsistent, depending on the order of the user jdk.table entries.

Highlighted also by the fact that ALLOW_DIFFERENT_JDK_VERSION
flag is true by default, making in most cases return the
first jdk.table entry for the user. The fact that the jdk.table
behaves as a stack doesn't help in this case.

To simplify all of this, we are going to use the Embedded JDK
keeping this consistency with the AGP required JDK version.
Due that both being shipped together in each studio version this will
guarantee consistency of the version used.

Soon, a new UI will allow users to configure their JDK
for new projects under File -> New Project Setup -> Project Structure

Bug: 273803862
Test: GradleSyncJdkIntegrationTest
Change-Id: I132faf674ab8725c37142fa01bcf739fcaa9afe7