[NPW,.kts] Don't apply kts conversion to the project name
While we still think of our build files as being generated in Groovy
and then translated through a carefully-crafted series of string
replacements to KotlinScript, we are vulnerable to converting things
that should not be converted.  In this instance, it is the content of
a string (which from the input constraints forbidding quote and
backslash, and escaping any dollar signs, is language-neutral).

Fixes: 273978558
Test: existing UI test
Change-Id: I340223c87daa2c2b974e6c81f8e139e6f313585c