Sync Download Info: ensure no events missed on init
There were two potential problems from concurrent nature of this UI init
process with the previous implementation. This is because event is
created and sent from Gradle connection thread, UI is created in some
background thread and then updates are rendered on EDT.
1) If build is ultra-short it can be that it is finished before UI is
initialised. That would result in an error with disposable registration.
2) In case there are any download event happening before UI is
initialised these events are missed.

To address these potential issues rearrange slightly the involved
components. Instead of having ModelRefresher (responsible for processing
updates from background thread to UI) being part of each UI Model that
is created later, introduce one DownloadInfoDataModel doing the same
but not attached to UI model. Instead it is bound to
DownloadsInfoPresentableEvent (which creates UI) and thus it is created
in right time (on build start) and in fact only one instance per build.
Then any created from that presentable event ui models should be
subscribed to this data model instance. On subscription it also provides
all download events that have been received already, so no data will be
missed on UI.

Bug: 271258614
Test: updated tests and added relevant test cases
Change-Id: Icc605ecaf000f6037496d27e7760ca7bbd514104