Add debugger task to RunStats metrics
RunStats allows recording metrics for various launch
tasks, but did not record "waitPreviousProcessTermination"
and more importantly "startDebuggerSession". The later
will be useful to compare failure rate of starting a debug session
with ddmlib vs adblib.

The way this is implemented is by re-using the "LaunchTask" entries
with "fake" IDs, so that we don't have to update proto and metrics.
There does not seem to be any real downside of re-using existing
entries.

Bug: N/A
Test: Manually tested metrics are logged
Change-Id: I8048492543e4b85d91d5e7b61979a8493fa225f0