Sync Download Info: remove unnecessary message layer
Refactoring on previous change makes it possible to get rid of
one of the middle layers, DownloadsInfoUIModelNotifier, and the way
it propagates updates further in the pipeline.
Instead, since now it is guaranteed only one DownloadInfoDataModel
instance per build is created, we can install it directly to gradle
events listener. Conversion logic from DownloadsInfoUIModelNotifier
will now be part of DownloadInfoDataModel.
We can not use message bus subscription to pass updates to ui models
because it does not provide 'replay' functionality to receive missed
data due to late subscription. We have to keep actual reference to
DownloadInfoDataModel for this anyway, meaning we can just subscribe
listener directly the way we already implemented in previous change.

Bug: 271258614
Test: updated existing tests
Change-Id: I92dab5af0bc77e28b053adf51dd8985c48ac6801