Add telemetry to usage searcher
This telemetry matches what exists on the V1 usage provider, so that we
can accurately compare telemetry. This includes starting timing *after*
a read lock is obtained, even though arguably that should be part of the
timing gathered. It also means that we don't log telemetry when there
are no usages found, even though we spent time looking for them.

Bug: 272736137
Test: DaggerCustomUsageSearcherTestV2
Change-Id: I72a76a104a1c44fe27acd4b0db94a5aa7306a5cd