Add version model to GradleVersionCatalogModel
Now calling addDependency via DependencyManager or
DependencyHelper will create catalog library declaration
with version as a reference.

VersionCatalogModel.versionDeclarations was added with
corresponding TOML specific version model. It allows to
read versions in all formats - literal, map, or part of
literal. Also it can create dependency with any version
format except in-place-map.

Version transformation from literal to map and back
was not implemented. We will need this case if,
for example, we add require and prefer attributes
to version that is stored as literal element.

Bug: 278094785
Test: GradleVersionCatalogLibrariesTest
Change-Id: Iecf59c9d8793febe2802669c2d97ab542b7ec683