Take dependency version into account on catalog search
When AutoImport insert some dependency to project with
catalog it can see existing dependency. If dependency
has the wrong version and it does not insert anything.
In case dependency with the wrong version is in catalog
- Auto Import does not see it and cannot make this
decision.

To make this happen DependencyManager look through
build dependencies and catalog declarations.
Now it searches in build dependencies with any version
and catalog dependency with particular version.

For this case we insert catalog declaration with proper
version and reference to dependency. This way module will
not have conflict but there can be a conflict
for other modules that use given one. Still, Gradle build
will be able to run and show conflict message.

It's up to lint checkers to verify such situations
and show warning in case user has same libraries
with different versions in catalog.

Fixed one side issue - DependencyManager resolves
dependency version in case it's dynamic based on
what other dependencies are there for module.
So even when adding declaration to catalog we need to
have module to resolve version.

Bug: 275602766
Test: GradleDependencyManagerTest
Change-Id: I1d87073e605ce018fc1202dd614b5edeb76f56c4