Add UI Check mode for Compose Preview
This creates a UI Check mode for the Compose Preview. This adds a new
button in the toolbar of each preview. Upon clicking, this takes the
preview and creates four versions of it with replacing the device with
one of the reference devices. It also runs ATF checks on each of these
previews.

Bug: 274150043
Test: Test added
Change-Id: I982c97cecdc53a499a3b22860bbc9ce4c16802b5