Add deprecation msg panel for Energy Profiler
The Power Profiler in the System Trace has replaced
the existing Energy Profiler. Thus, when the feature
flag for the Power Profiler is enabled, it will turn
on a deprecation message panel on the Energy Profiler
slice of the Studio Monitor Stage View, indicating that
the power data has moved.

Bug: 254116554
Test: manual
Change-Id: Ibeada48bce04e89e593455a7fa94d857e7bf595a