Add missing tooltip string
With ag/21685470, I added a missing string for Modules included in
Subcomponents. I didn't realize at the time that such a string also
required adding a tooltip for the gutter icon; the result is that
hovering over a module that's included in exactly 1 subcomponent and
nothing else will cause an error and no tooltip is shown. This change
fixes that.

This is not covered by tests; there aren't existing tests for the
"Module in Subcomponent" case, and since V1 of Dagger is quickly
becoming obsolete, I'm opting for manual verification of this fix.

Bug: 271336905
Test: Manually verified
Change-Id: I9bc3357edaae7fda9f671cec08206f691437ff2c