Work around minimum-size issue
This is a temporary workaround for b/260240804, a bug which causes the
minimum size of a dialog to be used on macOS due to a race condition in
the JDK. The workaround is to set the minimum size of the affected
dialogs to their preferred sizes. As a result, users will be unable to
shrink such dialogs.

This workaround should be removed once b/260240804 is properly fixed.

Bug: 260240804
Test: manually tested
Change-Id: I4fe414ebf08f975bc3e389ea5345c707dccc14c3