Disable all "unknown.sdk" registries to avoid displaying suggestions
The platform UnknownSdkTracker is strictly linked with provided
Project JDK under project-jdk-name on .idea/misc.xml, displaying a
banner message only on specific files like java or kt, when users
opened the project with an invalid jdk.table entry.
Those are some messages:
 - JDK "X" is missing (no jdk entry)
 - JDK "Y" is not found on the disc or corrupted (wrong jdk path)

This creates confusion for the user that just opened the project, but
at the same time is somehow conflicting with the recent logic we have
in place about updating the Project JDK after a successful sync.
Meaning that in any case, any user selection is going to be
overriding with the specified Gradle JDK for consistency reasons but
also to guarantee a valid JDK will be used to resolve project symbols

Bug: 236055498
Test: UnknownSdkTrackerTest
Change-Id: I878b9e409e69df491d4ed62b0cfa9ca1759e4a0b