Seperate battery track name and unit params
Before, only a track name was passed into models and views,
and a mapping from that specific track name was used to
identify the corresponding unit. This meant that if the track name
changed in the future, the unit would no longer be mappable.
This was fragile and prone to breaking with upcoming track name
mapping changes.

To address this, the track name and unit were seperated as
different parameters into passed in models and views. Now,
we can use the original raw name of the track to idenitify
the unit initially and not worry about the track name changing
in the future.

Bug: 254116554
Test: updated
Change-Id: I1879b1669d131e661e0c566d64d5f21ea9b17fd8