Make internal classes accessible for ScreenshotTesting
The following classes are needed by the stand alone screenshot
testing tool. These classes are loaded via the plugin.xml from
android studio, however the screenshot tool does not load a
plugin as such these classes cannot be instantiated via reflection

This change can be temporary as we refactor the standalone tool
to have fewer dependencies on AndroidStudio. To do this we need
to continue to break out functionality into interfaces and their
own modules.

Test: N/A
Bug: 277647080
Change-Id: Ie2a26934bcc5f28a75cd9c947b4041a150233d77