[Vitals] Increase page size for metrics and accessible apps call
...to 1000, which is also the default value
if not set.

Also, we remove the gRPC inbound message size limit
for now, in other words, now the limit is the default
limit: 4MB.

The above are to make sure we get as many data points
(rows) as we can in one call to get approximately
right counts when querying metrics.

Maybe in the future, either we have a simpler API or
we need to make a bunch of calls ourselves to get the
paginated data.

Bug: n/a
Test: n/a
Change-Id: I40c4d6665c78cfeda1c0c5d34458f8b49e31716a