Add read action in EditorStatsUtil
EditorStatsUtilTest.kotlinFileTypes is failing after the IDEA
231.4840.387 update due to a read access outside of a read action. In
some exploratory testing, this doesn't appear to happen in the actual
product (at least not consistently), likely because the information is
already cached in memory and a read access isn't always necessary.

Regardless, this call happens on a worker thread and it's safe to put it
in a read lock.

Fixes: 278937651
Test: EditorStatsUtilTest
Change-Id: I6059276917f98cc8f49152fe210f24185e1047e2