Break project modification out of core nav editor
At the moment the generic Project and Module System interfaces are
very Gradle-specific, and ASwB has to contort itself to implement
those interfaces (e.g. dependencies are expressed in terms of
GradleCoordinate; there are implicit assumptions about the division of
the project into Modules, etc.)  The experience of writing an IDE
feature which needs to modify the project is painful, working through
layers of (sometimes inappropriate) abstraction to achieve something
which does not necessarily feel natural even to Gradle users, and
definitely does not work well for ASwB.

This CL is an experiment to see whether an alternative approach is
useful.  The design constraints are at least:

- we don't want the Project System to own everything; it should not be
  the case that arbitrary feature authors need to go through the
  Project System team to do things;

- the generic Project System is still a useful interface, for querying
  things about the project or its components (at whatever level of
  granularity: Modules, VirtualFiles, individual code symbols);

- having a generic interface for project modification is unlikely to
  be useful, as we need to support either the intersection or the union
  of all available operations; the intersection is probably null for
  practical purposes, and the union implies having to support
  unsupportable operations in each concrete project system
  implementation.

- each concrete project system is in a good position to provide
  build-system-specific operations, but that only helps if the IDE
  feature can somehow access those features;

The approach in this change is to divide resposibilities further.  A
feature (in this prototypical case, the Navigation Editor) wishing to
perform a modification to the project defines an extension point.  It
also defines two new modules, one for Gradle and one for Blaze; those
modules are permitted to depend directly on the concrete build
system (e.g. not just on projectSystem but on projectSystem.gradle).
Those modules provide an implementation of the extension points, with
a small amount of type erasure punning allowing them to define methods
with the compile-time type of the concrete project system, which can
be called from the generic module with an argument of the correct
run-time type but a compile-time type of AndroidProjectSystem.

The implementations of those extension points can then do what they
like.  For now we simply copy the existing implementation, but we
expect to see them diverge fairly quickly to express operations that
are actually natural in each of the build systems.

In order to allow that divergence, we alter slightly the signatures of
the functions making modifications, preferring to keep the feature
domain data objects (e.g. a NavEditorSurface or a NlModel) rather than
Project System objects such as Project or Module.  In the process this
removes uses of Gradle-specific objects (e.g. GradleCoordinate) in the
generic feature module.  The behaviour of the modification functions
is unchanged in this CL.

The module in which each extension implementation point lives depends
on both the concrete project system module and the IDE feature
module.  The IDE feature module (or one of its transitive parents)
should have a runtime dependency on the implementation module, but the
extension implementation point module should otherwise not be depended
on by anything.

Bug: 279149611
Bug: 279886738
Test: existing
Change-Id: I9f1b150d3c9e0849d35f851a7a5ec3c6963f8351